requirements.txt
annotated-types==0.7.0
anyio==4.11.0
audioop-lts==0.2.2
certifi==2025.10.5
charset-normalizer==3.4.4
elevenlabs==2.21.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
keyboard==0.13.5
langdetect==1.0.9
numpy==2.3.4
PyAudio==0.2.14
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
python-dotenv==1.2.1
requests==2.32.5
setuptools==80.9.0
simpleaudio==1.0.4
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.3
standard-aifc==3.13.0
standard-chunk==3.13.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
webrtcvad==2.0.10
websockets==15.0.1
